refactor: alteração de organização da arquitetura hexagonal para clean architeture
