+++ src/main/java/br/fiap/projeto/contexto/identificacao/adapter/controller/IdentificacaoRestAdapterController.java
@@ -1,0 +1,4 @@
+package br.fiap.projeto.contexto.identificacao.adapter.controller;
+
+public class IdentificacaoRestAdapterController {
+}
