--- src/main/java/br/fiap/projeto/contexto/identificacao/application/rest/request/ClienteRequestDTO.java
@@ -1,43 +1,0 @@
-package br.fiap.projeto.contexto.identificacao.application.rest.request;
-
-public class ClienteRequestDTO {
-
-    private String nome;
-
-    private String cpf;
-
-    private String email;
-
-    public ClienteRequestDTO() {
-    }
-
-    public ClienteRequestDTO(String nome, String cpf, String email) {
-        this.nome = nome;
-        this.cpf = cpf;
-        this.email = email;
-    }
-
-    public String getNome() {
-        return nome;
-    }
-
-    public void setNome(String nome) {
-        this.nome = nome;
-    }
-
-    public String getCpf() {
-        return cpf;
-    }
-
-    public void setCpf(String cpf) {
-        this.cpf = cpf;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-}
