+++ src/main/java/br/fiap/projeto/contexto/identificacao/usecase/exception/EntidadeNaoEncontradaException.java
@@ -1,0 +1,9 @@
+package br.fiap.projeto.contexto.identificacao.usecase.exception;
+
+public class EntidadeNaoEncontradaException extends BaseException {
+
+    public EntidadeNaoEncontradaException(String message) {
+        super(4001, message);
+    }
+
+}
