+++ src/main/java/br/fiap/projeto/contexto/identificacao/usecase/exception/EntradaInvalidaException.java
@@ -1,0 +1,8 @@
+package br.fiap.projeto.contexto.identificacao.usecase.exception;
+
+public class EntradaInvalidaException extends BaseException {
+
+    public EntradaInvalidaException(String message) {
+        super(4002, message);
+    }
+}
