+++ src/test/java/br/fiap/projeto/contexto/identificacao/ClienteValidacoesTest.java
@@ -1,9 +1,9 @@
 package br.fiap.projeto.contexto.identificacao;
 
-import br.fiap.projeto.contexto.identificacao.domain.entity.Cliente;
-import br.fiap.projeto.contexto.identificacao.domain.vo.Cpf;
-import br.fiap.projeto.contexto.identificacao.domain.vo.Email;
-import br.fiap.projeto.contexto.identificacao.domain.exception.EntradaInvalidaException;
+import br.fiap.projeto.contexto.identificacao.entity.entity.Cliente;
+import br.fiap.projeto.contexto.identificacao.entity.vo.Cpf;
+import br.fiap.projeto.contexto.identificacao.entity.vo.Email;
+import br.fiap.projeto.contexto.identificacao.usecase.exception.EntradaInvalidaException;
 import org.junit.jupiter.api.Test;
 
 import java.util.UUID;
